backend/pyproject.toml
[tool.poetry]
name = "chefstream-backend"
version = "0.1.0"
description = "AI-powered recipe extractor backend"
authors = ["ChefStream Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.5.0"
alembic = "^1.14.0"
httpx = "^0.27.0"
python-multipart = "^0.0.12"
youtube-transcript-api = "^0.6.2"
yt-dlp = "^2025.2.22"
google-genai = "^1.0.0"
webvtt-py = "^0.4.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^8.0.0"}
ua-parser = "^0.18.0"
bcrypt = "^4.2.0"
geoip2 = "^4.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = "tests.py test_*.py *_tests.py"

[tool.coverage.run]
omit = [
    "app/services/security_service.py",
    "app/services/email_service.py",
    "app/services/geoip.py",
    "app/api/endpoints/security.py",
    "app/services/oauth.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = false  # Too strict for gradual typing
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing: start permissive
check_untyped_defs = false  # Don't check untyped functions yet
warn_redundant_casts = true
warn_unused_ignores = false  # Avoid noise during gradual typing
warn_no_return = true
warn_unreachable = true
strict_optional = false  # Too strict for gradual typing
ignore_missing_imports = true  # Ignore missing stubs for now

# Exclude patterns
exclude = [
    "^tests/",
    "^migrations/",
    "^alembic/",
]

# Per-module options for gradual typing
# These modules will have strict typing enforced
[[tool.mypy.overrides]]
module = "app.services.security_service"
disallow_untyped_defs = false  # Will enable in Phase 3
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.email_service"
disallow_untyped_defs = false  # Will enable in Phase 3
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.geoip"
disallow_untyped_defs = false  # Will enable in Phase 3
ignore_missing_imports = true
